requirements.txt


python-dotenv

#pandas

#plotly

#openai # 1.3.8
#langchain # 0.0.348
#tiktoken
#faiss-cpu

#langchain
#einops
#accelerate

langchain
langchain-community # for hugging face hub
huggingface_hub

#pytest
